telegram v0.5.1 Telegram.Bot behaviour

A simple BOT behaviour and DSL.

Example

defmodule Simple.Bot do
  use Telegram.Bot,
    token: "123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11",
    username: "simple_bot",
    auth: ["user1", "user2"]

  command ["ciao", "hello"], args do
    # handle the commands: "/ciao" and "/hello"

    # reply with a text message
    request("sendMessage",
      chat_id: update["chat"]["id"],
      text: "ciao! #{inspect args}")
  end

  command unknown do
    request("sendMessage", chat_id: update["chat"]["id"],
      text: "Unknow command `#{unknown}`")
  end

  message do
    request("sendMessage", chat_id: update["chat"]["id"],
      text: "Hey! You sent me a message: #{inspect update}")
  end

  edited_message do
    # handler code
  end

  channel_post do
    # handler code
  end

  edited_channel_post do
    # handler code
  end

  inline_query _query do
    # handler code
  end

  chosen_inline_result _query do
    # handler code
  end

  callback_query do
    # handler code
  end

  shipping_query do
    # handler code
  end

  pre_checkout_query do
    # handler code
  end

  any do
    # handler code
  end
end

See Telegram.Bot.Dsl documentation for all available macros.

Options

use Telegram.Bot,
  token: "your bot auth token",   # required
  username: "your bot username",  # required
  auth: ["user1", "user2"],       # optional, list of authorized users
                                  # or authorizing function (String.t -> boolean)
  purge: boolean()                # purge old messages at startup, default: false
  restart: policy                 # optional, default :permanent

Execution model

The bot defined using the Telegram.Bot behaviour is based on Task and will run in a single erlang process, processing updates sequentially.

You can add the bot to you application supervisor tree, for example:

  children = [Simple.Bot, ...]
  opts = [strategy: :one_for_one, name: MyApplication.Supervisor]
  Supervisor.start_link(children, opts)

or directly start and link the bot with:

{:ok, pid} = Simple.Bot.start_link()

Functions

process_update(update, acc, context)

Callbacks

handle_auth(username)

handle_auth(username :: String.t()) :: boolean()

handle_update(token, update)

handle_update(token :: String.t(), update :: map()) :: any()

init()

init() :: any()