Telegram.Bot.Async.Supervisor (telegram v0.13.0)

Bot Supervisor - Asynchronous update dispatching

The Bot Telegram.Bot.handle_update/2 function is called a dynamically spawned Task, so every update is handled by an isolated Task process. (this can be controlled/limited with the max_bot_concurrency option)

Types

options()

@type options() :: [
  bot_behaviour_mod: module(),
  token: Telegram.Types.token(),
  max_bot_concurrency: Telegram.Types.max_bot_concurrency()
]

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(opts)

@spec start_link(options()) :: Supervisor.on_start()