telegram v0.5.1 Telegram.Bot.Dsl

Telegram.Bot DSL macros.

Update object

Every macro inject in the body scope a update variable holding the received specific update object (Message, InlineQuery, ...), with the exception of any/1 where update holds the Update object.

Reference: Update

Functions

any(list) (macro)

Match any Telegram update. This should be the very last fallback clause in your bot.

any do
  IO.puts("ANY #{inspect update}")
end

callback_query(list) (macro)

Match a generic Telegram callback_query.

callback_query do
  # handler code
end

channel_post(list) (macro)

Match a generic Telegram channel_post.

channel_post do
  # handler code
end

chosen_inline_result(query, list) (macro)

Match a Telegram chosen_inline_query

chosen_inline_result query do
  # handler code
end

command(text, list) (macro)

Match any other not matched command.

command "start", args do
  # handler code
end

command unknown do
  # ex: telegram -> "/end hello"
  #     unknown = "end hello"
end

command(text, args, list) (macro)

Match Telegram "/command arg1 arg2" (with args, if any).

command "start", args do
  # ex: telegram -> "/start hello 1"
  #     args = ["hello", "1"]
end

command "start", ["hello", "2"] do
  # ex: telegram -> "/start hello 2"
  #
  # NOTE: the args are not whitespace normalized (stripped)
  #       so the pattern matching is an identity match with the string
  #       "/start hello 2" (one space char separator),
  #       ie. "/start   hello 2" won't match
end

or a set of commands:

command ["start", "go"], args do
  # handler code
end

edited_channel_post(list) (macro)

Match a generic Telegram edited_channel_post.

edited_channel_post do
  # handler code
end

edited_message(list) (macro)

Match a generic Telegram edited_message.

edited_message do
  # handler code
end

halt(message, system_stop \\ false)

halt(message :: String.t(), system_stop :: boolean()) :: no_return()

Halts the Bot.

command "stop_bot", _args do
  halt("user requested to stop the bot")
end

command "stop_system", _args do
  halt("user requested to stop the system", true)
end

inline_query(query, list) (macro)

Match a Telegram inline_query

inline_query query do
  # ex: telegram -> "@your_bot queeery!"
  #     query = "queeery!"
end

message(list) (macro)

Match a generic Telegram message.

message do
  # ex: telegram -> "go!"
  #     update["text"] = "go!"
end

Note:

Since Telegram commands are just messages starting with "/", this can match also commands. In general you should put command/3 before message/2 clause:

command "start", args, do: :something
command unknown, do: :something_for_unknown_commands
message do: :something_for_messages

pre_checkout_query(list) (macro)

Match a generic Telegram pre_checkout_query.

pre_checkout_query do
  # handler code
end

request(method, options) (macro)

A small wrapper for Telegram.Api.request. Autofill the token parameter with the Telegram.Bot configured.

command "hello", args do
  request "sendMessage", chat_id: update["chat"]["id"],
    text: "Hello #{inspect update["from"]["username"]}"
end

shipping_query(list) (macro)

Match a generic Telegram shipping_query.

shipping_query do
  # handler code
end