Telegram.ChatBot behaviour (telegram v1.1.0)

Telegram Chat Bot behaviour.

The difference with Telegram.Bot behaviour is that the Telegram.ChatBot is "statefull" per chat_id, (see chat_state argument).

Given that every "conversation" is associated with a long running process is up to you to consider a session timeout in your bot state machine design. If you don't you will saturate the max_bot_concurrency capacity and then your bot won't accept any new conversation. For this you can leverage the underlying gen_server timeout including the timeout in the return value of the init/1 or handle_update/3 callbacks or, if you need a more complex behaviour, via explicit timers in you bot.

Example

defmodule HelloBot do
  use Telegram.ChatBot

  @session_ttl 60 * 1_000

  @impl Telegram.ChatBot
  def init(_chat) do
    count_state = 0
    {:ok, count_state, @session_ttl}
  end

  @impl Telegram.ChatBot
  def handle_update(%{"message" => %{"chat" => %{"id" => chat_id}}}, token, count_state) do
    count_state = count_state + 1

    Telegram.Api.request(token, "sendMessage",
      chat_id: chat_id,
      text: "Hey! You sent me #{count_state} messages"
    )

    {:ok, count_state, @session_ttl}
  end

  def handle_update(update, _token, count_state) do
    # ignore unknown updates

    {:ok, count_state, @session_ttl}
  end

  @impl Telegram.ChatBot
  def handle_info(msg, _token, _chat_id, count_state) do
    # direct message processing

    {:ok, count_state}
  end

  @impl Telegram.ChatBot
  def handle_timeout(token, chat_id, count_state) do
    Telegram.Api.request(token, "sendMessage",
      chat_id: chat_id,
      text: "See you!"
    )

    {:stop, count_state}
  end
end

Types

chat()

@type chat() :: map()

chat_state()

@type chat_state() :: any()

t()

@type t() :: module()

Callbacks

handle_info(msg, token, chat_id, chat_state)
(optional)

@callback handle_info(
  msg :: any(),
  token :: Telegram.Types.token(),
  chat_id :: String.t(),
  chat_state :: chat_state()
) ::
  {:ok, next_chat_state :: chat_state()}
  | {:ok, next_chat_state :: chat_state(), timeout :: timeout()}
  | {:stop, next_chat_state :: chat_state()}

On handle_info callback.

Can be used to implement bots that act on scheduled events (using Process.send/3 and Process.send_after/4) or to interact via direct message to a a specific chat session (using lookup/2).

This callback is optional. If one is not implemented, the received message will be logged.

handle_timeout(token, chat_id, chat_state)
(optional)

@callback handle_timeout(
  token :: Telegram.Types.token(),
  chat_id :: String.t(),
  chat_state :: chat_state()
) ::
  {:ok, next_chat_state :: chat_state()}
  | {:ok, next_chat_state :: chat_state(), timeout :: timeout()}
  | {:stop, next_chat_state :: chat_state()}

On timeout callback.

This callback is optional. A default implementation is injected with "use Telegram.ChatBot", it just stops the bot.

handle_update(update, token, chat_state)

@callback handle_update(
  update :: Telegram.Types.update(),
  token :: Telegram.Types.token(),
  chat_state :: chat_state()
) ::
  {:ok, next_chat_state :: chat_state()}
  | {:ok, next_chat_state :: chat_state(), timeout :: timeout()}
  | {:stop, next_chat_state :: chat_state()}

Receives the telegram update event and the "current" chat_state. Return the "updated" chat_state.

init(chat)

@callback init(chat :: chat()) ::
  {:ok, initial_state :: chat_state()}
  | {:ok, initial_state :: chat_state(), timeout :: timeout()}

Invoked once when the chat starts. Return the initial chat_state.

Functions

lookup(token, chat_id)

@spec lookup(Telegram.Types.token(), String.t()) ::
  {:error, :not_found} | {:ok, pid()}

Lookups the pid of a specific chat session.

It is up to the user to define and keep a mapping between the business logic specific session identifier and the telegram chat_id.