Telegram.Webhook (telegram v1.2.3)

Telegram Webhook supervisor.

This modules can the used to start a webserver exposing a webhoook endpoint where the telegram server can push updates for your BOT. On start the webhook address for the BOT is posted to the telegram server via the setWebHook method.

Usage

WebServer adapter

Two Plug compatible webserver are supported:

Telegram.WebServer.Bandit (default): use Bandit

Telegram.WebServer.Cowboy: use Plug.Cowboy

You should configure the desired webserver adapter in you app configuration:

config :telegram,
  webserver: Telegram.WebServer.Bandit

# OR

config :telegram,
  webserver: Telegram.WebServer.Cowboy

and include in you dependencies one of:

{:plug_cowboy, "~> 2.5"}

# OR

{:bandit, "~> 1.0"}

Supervision tree

In you app supervision tree:

webhook_config = [
  host: "myapp.public-domain.com",
  port: 443,
  local_port: 4_000
]

bot_config = [
  token: Application.fetch_env!(:my_app, :token_counter_bot),
  max_bot_concurrency: Application.fetch_env!(:my_app, :max_bot_concurrency)
]

children = [
  {Telegram.Webhook, config: webhook_config, bots: [{MyApp.Bot, bot_config}]}
  ...
]

opts = [strategy: :one_for_one, name: MyApp.Supervisor]
Supervisor.start_link(children, opts)

Ref

https://core.telegram.org/bots/api#setwebhook

https://core.telegram.org/bots/webhooks

Direct Phoenix / Plug integration

If want to integrate the webhook in a Phoenix / Plug based application facing internet, you should (for each BOT):

add a POST route in your endpoints to receive the webhook updates and dispatch them to the BOT.

active the webhook mode posting a setWebHook request to the telegram server.

Refer to this module implementation as a guide line for the above.

Types

config()

@type config() :: [
  host: String.t(),
  port: :inet.port_number(),
  local_port: :inet.port_number(),
  max_connections: 1..100,
  set_webhook: boolean()
]

Webhook configuration.

host: (reverse proxy) hostname of the HTTPS webhook url (required)

port: (reverse proxy) port of the HTTPS webhook url (optional, default: 443)

local_port: (backend) port of the application HTTP web server (optional, default: 4000)

max_connections: maximum allowed number of simultaneous connections to the webhook for update delivery (optional, defaults 40)

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(list)

@spec start_link(config: config(), bots: [Telegram.Types.bot_spec()]) ::
  Supervisor.on_start()